+++ .env.example
@@ -1,3 +1,4 @@
 DATABASE_URL="mysql://johndoe:randompassword@localhost:3306/mydb"
 EMAIL_USER=email_address
 EMAIL_PASSWORD=password
+URL_EMAIL_VERIFY=http://localhost:5173
