+++ src/lib/components/index.ts
@@ -1,7 +1,7 @@
 // place files you want to import through the `$lib` alias in this folder.
 export { default as Main } from "./main.svelte";
 export { default as Footer } from "./footer.svelte";
-export { default as Navigation } from "./navigation.svelte";
+export { default as MainNav } from "./main-nav.svelte";
 export { default as Hero } from "./hero.svelte";
 export { default as Founders } from "./founders.svelte";
 export { default as Videos } from "./video.svelte";
@@ -9,3 +9,4 @@ export { default as Join } from "./join.svelte";
 export { default as Register } from "./register.svelte";
 export { default as Login } from "./login.svelte";
 export { default as MissionVision } from "./mission-vision.svelte";
+export { default as UserNav } from "./user-nav.svelte";
