+++ src/lib/components/main-nav.svelte
