+++ src/lib/server/email.service.ts
@@ -30,6 +30,32 @@ export const getEmailVerificationCodeByUserId = async (userId: string) => {
   return db.emailVerificationCode.findUnique({ where: { userId } });
 };
 
+export const updateEmailVerified = async (
+  user: Omit<
+    User,
+    "id" | "username" | "email" | "firstName" | "lastName" | "hashed_password"
+  >,
+  id: string,
+): Promise<User> => {
+  const { email_verified } = user;
+
+  return db.user.update({
+    where: { id },
+    data: {
+      email_verified,
+    },
+    select: {
+      id: true,
+      username: true,
+      email: true,
+      firstName: true,
+      lastName: true,
+      hashed_password: true,
+      email_verified: true,
+    },
+  });
+};
+
 export const deleteEmailVerificationCodeByUserId = async (userId: string) => {
   return db.emailVerificationCode.delete({ where: { userId } });
 };
